requirements.txt
streamlit
opencv-python
numpy
deepface
Pillow
opencv-contrib-python

